+++ CHANGELOG.md
@@ -8,6 +8,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/),
 and this project adheres to [Semantic Versioning](https://semver.org/).
 
 ## [Unreleased]
+### Added
+- New function `search.values()` to return an array of all unique values of a key [#191].
+
 ### Fixed
 - Updated `markdown_it`, `liquid`, `std`, `esbuild`, `postcss_autoprefixer` and `pug`.
 
@@ -1493,6 +1496,7 @@ The first version.
 [#175]: https://github.com/lumeland/lume/issues/175
 [#181]: https://github.com/lumeland/lume/issues/181
 [#182]: https://github.com/lumeland/lume/issues/182
+[#191]: https://github.com/lumeland/lume/issues/191
 
 [Unreleased]: https://github.com/lumeland/lume/compare/v1.7.4...HEAD
 [1.7.4]: https://github.com/lumeland/lume/compare/v1.7.3...v1.7.4